MCOL-4010 - fixes compilation errors on x64 w/-Werror
Merged in Sergei's patch.

+++ storage-manager/src/WriteTask.cpp
@@ -74,7 +74,7 @@ bool WriteTask::run()
 
     while (readCount < cmd->count)
     {
-        uint toRead = min(cmd->count - readCount, bufsize);
+        uint toRead = min(static_cast<uint>(cmd->count - readCount), bufsize);
         success = read(&databuf[0], toRead);
         check_error("WriteTask read data", false);
         if (success==0)
